refactor: move tests to skeleton.rs
